refactor(api): migrate to user-centric relations and implement RBAC

+++ src/controller/commentController.ts
@@ -6,6 +6,8 @@ import {
   deleteComment,
   getCommentById,
 } from "../models/commentModel.js";
+import { getUserById } from "../models/userModel.js";
+import { getPostById } from "../models/postModel.js";
 
 export const getAllCommentController = async (req: Request, res: Response) => {
   try {
@@ -20,7 +22,7 @@ export const getAllCommentController = async (req: Request, res: Response) => {
 
 export const createCommentController = async (req: Request, res: Response) => {
   try {
-    const newItem = await createComment(req.body);
+    const newItem = await createComment(req.body, req.user as number);
     res.status(201).json(newItem);
   } catch (error) {
     console.error("Error fetching posts:", error);
@@ -30,10 +32,17 @@ export const createCommentController = async (req: Request, res: Response) => {
 export const updateCommentController = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
-    const updatedItem = await updateComment(Number(id), req.body);
-    if (!updatedItem) {
-      res.status(404).json({ message: "Comment not found" });
+    const item = await getCommentById(Number(id));
+    if (!item) {
+      return res.status(404).json({ message: "Comment not found" });
     }
+    const requestUser = await getUserById(req.user as number);
+    if (requestUser?.id !== item.user_id) {
+      return res
+        .status(403)
+        .json({ message: " You are not authorized to update this comment" });
+    }
+    const updatedItem = await updateComment(Number(id), req.body);
     res.json(updatedItem);
   } catch (error) {
     console.error("Error fetching posts:", error);
@@ -43,6 +52,27 @@ export const updateCommentController = async (req: Request, res: Response) => {
 export const deleteCommentController = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const item = await getCommentById(Number(id));
+    if (!item) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+
+    const post = await getPostById(item.post_id);
+    if (!post) {
+      res.status(404).json({ message: "Post not found" });
+    }
+
+    const requestUser = await getUserById(req.user as number);
+    if (
+      requestUser?.id !== item.user_id &&
+      requestUser?.role !== "ADMIN" &&
+      requestUser?.role !== "MODERATOR"
+      && post?.user_id !== requestUser?.id
+    ) {
+      return res
+        .status(403)
+        .json({ message: " You are not authorized to update this comment" });
+    }
     const deletedItem = await deleteComment(Number(id));
     if (!deletedItem) {
       res.status(404).json({ message: "Comment not found" });
